Add makefile Switch GPS to "always on" mode. Use SPI instead of I2C for display, for much faster updates. Correctly detect usb/bluetooth and switch to the right port

+++ rtc.cpp
@@ -43,7 +43,7 @@ time_t RTC::get_time_t() {
     get_day_of_week(rtc),
     rtc->day(),
     rtc->month(),
-    y2kYearToTm(rtc->year()),
+    static_cast<uint8_t>(y2kYearToTm(rtc->year())),
   };
   /*
   Serial.print(rtc->day()); Serial.print("/");
